Added memory usage to status view
