+++ plaso/multi_processing/task_engine.py
@@ -536,10 +536,12 @@ class TaskMultiProcessEngine(engine.MultiProcessEngine):
       for pid in list(self._process_information_per_pid.keys()):
         self._CheckStatusWorkerProcess(pid)
 
+      used_memory = self._process_information.GetUsedMemory()
+
       display_name = getattr(self._merge_task, u'identifier', u'')
 
       self._processing_status.UpdateForemanStatus(
-          self._name, self._status, self._pid, 0, display_name,
+          self._name, self._status, self._pid, used_memory, display_name,
           self._number_of_consumed_sources, self._number_of_produced_sources,
           self._number_of_consumed_events, self._number_of_produced_events,
           self._number_of_consumed_event_tags,
@@ -671,10 +673,8 @@ class TaskMultiProcessEngine(engine.MultiProcessEngine):
               u'the timeout period.').format(process.name, pid))
           processing_status = definitions.PROCESSING_STATUS_NOT_RESPONDING
 
-    process = self._processes_per_pid[pid]
     process_information = self._process_information_per_pid[pid]
-    memory_info = process_information.GetMemoryInformation()
-    used_memory = memory_info.data + memory_info.shared
+    used_memory = process_information.GetUsedMemory()
 
     self._processing_status.UpdateWorkerStatus(
         process.name, processing_status, pid, used_memory, display_name,
